Bunch of enhancements for scratchpad
Use wids instead of pids to save/restore states
We can use an arbitrary matcher instead of PIDs for scratchpad matching.
We can hide all scratchpads if needed in a group
We can set a scratchpad group to single so that only one of the
scratchpad is displayed at a given time.
We have a WindowTranform config that can be used to place the scratchpad
relative to the screen, which provides more capabilities.